# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Jatek Food Delivery API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /restaurants:
    get:
      operationId: listRestaurants
      tags:
        - restaurants
      summary: List restaurants
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: isLocal
          in: query
          schema:
            type: boolean
        - name: isOpen
          in: query
          schema:
            type: boolean
        - name: businessType
          in: query
          schema:
            type: string
        - name: ownerId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of restaurants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Restaurant"
    post:
      operationId: createRestaurant
      tags:
        - restaurants
      summary: Create a restaurant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRestaurantBody"
      responses:
        "201":
          description: Restaurant created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restaurant"
      description: "RBAC: admin only."
  /restaurants/{id}:
    get:
      operationId: getRestaurant
      tags:
        - restaurants
      summary: Get a restaurant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Restaurant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restaurant"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateRestaurant
      tags:
        - restaurants
      summary: Update a restaurant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRestaurantBody"
      responses:
        "200":
          description: Updated restaurant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restaurant"
      description: "RBAC: admin or owning restaurant_owner."
    delete:
      operationId: deleteRestaurant
      tags:
        - restaurants
      summary: Delete a restaurant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
      description: "RBAC: admin only."
  /restaurants/{id}/complete-profile:
    post:
      operationId: completeRestaurantProfile
      tags:
        - restaurants
      summary: Complete a restaurant business profile
      description: "RBAC: admin or the owning restaurant_owner. Required before accepting orders."
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteRestaurantProfileBody"
      responses:
        "200":
          description: Restaurant profile completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restaurant"
        "400":
          description: Invalid profile data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Restaurant not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /restaurants/{id}/stats:
    get:
      operationId: getRestaurantStats
      tags:
        - restaurants
      summary: Get restaurant statistics
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Restaurant statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestaurantStats"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Restaurant:
      type: object
      properties:
        id:
          type: integer
        ownerId:
          type: integer
        name:
          type: string
        description:
          type:
            - string
            - "null"
        address:
          type: string
        phone:
          type:
            - string
            - "null"
        imageUrl:
          type:
            - string
            - "null"
        coverImageUrl:
          type:
            - string
            - "null"
        logoUrl:
          type:
            - string
            - "null"
        category:
          type: string
        businessType:
          type: string
        isLocal:
          type: boolean
        isOpen:
          type: boolean
        deliveryTime:
          type:
            - integer
            - "null"
        deliveryFee:
          type:
            - number
            - "null"
        minimumOrder:
          type:
            - number
            - "null"
        rating:
          type:
            - number
            - "null"
        reviewCount:
          type: integer
        isVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
        legalName:
          type:
            - string
            - "null"
        ice:
          type:
            - string
            - "null"
        printerEmail:
          type:
            - string
            - "null"
        profileCompletedAt:
          type:
            - string
            - "null"
          format: date-time
        freeDeliveryThreshold:
          type: number
          description: Defaulted by the API when not stored on the restaurant row.
      required:
        - id
        - ownerId
        - name
        - address
        - category
        - businessType
        - isLocal
        - isOpen
        - reviewCount
        - isVerified
        - createdAt
    CreateRestaurantBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        address:
          type: string
        phone:
          type: string
        imageUrl:
          type: string
        coverImageUrl:
          type: string
        logoUrl:
          type:
            - string
            - "null"
        category:
          type: string
        businessType:
          type: string
        isLocal:
          type: boolean
        deliveryTime:
          type: integer
        deliveryFee:
          type: number
        minimumOrder:
          type: number
      required:
        - name
        - address
        - category
    UpdateRestaurantBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        address:
          type: string
        phone:
          type: string
        imageUrl:
          type: string
        coverImageUrl:
          type: string
        logoUrl:
          type:
            - string
            - "null"
        category:
          type: string
        businessType:
          type: string
        isLocal:
          type: boolean
        isOpen:
          type: boolean
        deliveryTime:
          type: integer
        deliveryFee:
          type: number
        minimumOrder:
          type: number
        isVerified:
          type: boolean
    RestaurantStats:
      type: object
      properties:
        totalOrders:
          type: integer
        completedOrders:
          type: integer
        totalRevenue:
          type: number
        averageRating:
          type:
            - number
            - "null"
        totalReviews:
          type: integer
        pendingOrders:
          type: integer
      required:
        - totalOrders
        - completedOrders
        - totalRevenue
        - totalReviews
        - pendingOrders
...
    CompleteRestaurantProfileBody:
      type: object
      required:
        - legalName
        - ice
      properties:
        legalName:
          type: string
          minLength: 2
        ice:
          type: string
          pattern: ^\d{8,15}$
        printerEmail:
          type:
            - string
            - "null"
          format: email
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT returned by /auth/login, /auth/verify-otp, or /backend/login. Some HTML document endpoints also accept the same token as a token query parameter.